backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
python-decouple==3.8
Pillow==11.0.0
django-filter==24.3
